chg: dev: Refactor for pages
